+++ barretenberg/cpp/src/barretenberg/stdlib/CMakeLists.txt
@@ -6,7 +6,6 @@ add_subdirectory(eccvm_verifier)
 add_subdirectory(goblin_verifier)
 add_subdirectory(merge_verifier)
 add_subdirectory(primitives)
-add_subdirectory(hypernova)
 add_subdirectory(honk_verifier)
 add_subdirectory(proof)
 add_subdirectory(translator_vm_verifier)
