--- barretenberg/cpp/src/barretenberg/stdlib/hypernova/CMakeLists.txt
@@ -1,1 +1,0 @@
-barretenberg_module(stdlib_hypernova hypernova stdlib_honk_verifier)
